Improvements to default function generation
* move default function generation to its own file
* also, areOverloadsPresentInDefiningScope, call isMethod
  on uAST node instead of UntypedFnSignature
* also, added isCompilerGenerated field to UntypedFnSignature

---
