+++ compiler/dyno/lib/resolution/Makefile.include
@@ -23,6 +23,7 @@ ALL_SRCS += dyno/lib/resolution/*.cpp
 DYNO_RESOLUTION_SRCS =                              \
            Resolver.cpp                             \
            can-pass.cpp                             \
+           default-functions.cpp                    \
            disambiguation.cpp                       \
            intents.cpp                              \
            prims.cpp                                \
