+++ compiler/dyno/lib/resolution/default-functions.h
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2021-2022 Hewlett Packard Enterprise Development LP
+ * Other additional copyright holders may be indicated within.
+ *
+ * The entirety of this work is licensed under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef DEFAULT_FUNCTIONS_
+#define DEFAULT_FUNCTIONS_
+
+#include "chpl/resolution/resolution-types.h"
+
+namespace chpl {
+namespace types {
+  class CompositeType;
+  class Type;
+}
+namespace resolution {
+
+
+TypedFnSignature* generateFieldAccessor(Context* context, UniqueString name,
+                                        const types::CompositeType* compType);
+
+/**
+  Given a type and a UniqueString representing the name of a method,
+  determine if the type needs a method with such a name to be
+  generated for it.
+*/
+bool needCompilerGeneratedMethod(Context* context, const types::Type* type,
+                                 UniqueString name, bool parenless);
+
+/**
+  Given a type and a UniqueString representing the name of a method,
+  determine if the type needs a method with such a name to be
+  generated for it, and if so, generates and returns a
+  TypedFnSignature representing the generated method.
+
+  If no method was generated, returns nullptr.
+*/
+const TypedFnSignature*
+getCompilerGeneratedMethod(Context* context, const types::Type* type,
+                           UniqueString name, bool parenless);
+
+
+} // end namespace resolution
+} // end namespace chpl
+
+#endif
